Move more stuff into SSBaseComboBox.
